fix(app): seal shared input sync paths
